refactor : Lottos 객체 제거하고 내부 기능을 Ranks로 이관
