+++ src/main/kotlin/com/nextstep/jngcii/lotto/Application.kt
@@ -4,7 +4,7 @@ import com.nextstep.jngcii.lotto.model.BonusNumber
 import com.nextstep.jngcii.lotto.model.Lotto
 import com.nextstep.jngcii.lotto.model.LottoMachine
 import com.nextstep.jngcii.lotto.model.LottoNumber
-import com.nextstep.jngcii.lotto.model.Lottos
+import com.nextstep.jngcii.lotto.model.Ranks
 import com.nextstep.jngcii.lotto.view.InputView
 import com.nextstep.jngcii.lotto.view.ResultView
 
@@ -14,13 +14,11 @@ fun main() {
     val lottos = LottoMachine.get(count)
     ResultView.printList(lottos)
 
-    val record = Lottos(lottos)
-
     val lastWeekNumbers = InputView.getNumbers { readLine() }
     val lastWeekLotto = Lotto(lastWeekNumbers.map { LottoNumber(it) })
     val number = InputView.getNumber { readLine() }
     val bonus = BonusNumber(number, lastWeekLotto)
 
-    val ranks = record.getResult(lastWeekLotto, bonus)
+    val ranks = Ranks(lottos, lastWeekLotto, bonus)
     ResultView.printResult(count, ranks)
 }
