+++ src/main/kotlin/com/nextstep/jngcii/lotto/model/Lottos.kt
@@ -1,9 +1,3 @@
 package com.nextstep.jngcii.lotto.model
 
-class Lottos(private val lottos: List<Lotto>) {
-    fun getResult(lastWeekLotto: Lotto, bonusNumber: BonusNumber) = lottos.mapNotNull {
-        val sameCount = it.getSameCount(lastWeekLotto)
-        val bonusMatch = it.contains(bonusNumber)
-        Rank.of(sameCount, bonusMatch)
-    }.let { Ranks(it) }
-}
+class Lottos(private val lottos: List<Lotto>)
